auto: issue #41 Extract SSE parser + typed client; thin frontend/src/lib/api.ts (1,216 lines)
